Complete basic print/println functionality for UART serial in TextOutput
Also, this commit removes the `TextOutput_debugMsg()` function from
`TextOutput` because this is not a core component of a text output
system (could be added back after discussion).

Also, this commit cleans up the UART initialisation code copied from the
VLE.

The only remaining functionality needed for mini-project-1 is interrupt
timers.