Basic keypad functionality complete
This commit completes the implementation of `Keypad_isKeyDown()`
which returns true or false for a given key defined in `Constants.h`.

A potential improvement would be storing and updating the state of
the keypad in one go, so that we could check whether multiple keys
were down at the same time.