Add flag recog work, LCD menu utils, core reqs
This commit adds a framework for developing recognition systems for
flags. Different flag recognition methods all involve an error
measurement between something on the platform, and some stored data for
each flag. This stored data may change for each method (critical
colour-checking points, or image histograms, or average colours etc).

Each different flag recognition method implements its own

    `uint32_t RecogMethod_errorFunction(FlagID flagId)`

function that takes in the index of one of the flags we need to
recognise, and returns an error value describing how far off
the contents of the platform is to the flag with this id, using
its own method. This function should produce an ordering over all flags,
so that we end up with one that is the least-wrong, with the lowest
error. Each recognition method stores its own precomputed/scanned data
for each flag. This data is what is indexed with flagID.

This function is then passed to the `FlagRecogniser` in `C1` which does
all the work of allowing the user to press '#' to start the
recognition process. It calls this `errorFunction` for each flag and
displays the closest match on the `LCDDisplay`.

An `LCDMenu` module has been added which makes the implementation of
any user options very easy. Options are added with names and function
callbacks and the module does the rest to allow the user to cycle
through the different options and select them. Selecting an option runs
the attached callback.

Work has been done to automatically detect the edges of the platform
using `ColourSensor` readings, as per requirement A2. These can be
found in `Motion_localisePlatform()`. This functionality is not yet
complete.

Requirement `B3` has been implemented but not yet tested. It finds
the highest value of a user-selected colour component (R/G/B), and makes
use of `LCDMenu`.

Odd lines throughout the `mbed` component of the project have been
modified to remove compiler errors. These were mainly unused variables,
and missing header #includes.